RequestColour is now implimented in the renderer and not in the skin Added way to overide VerticalProgressBar's colour (usefull for the graph) Will, at some point, be expanded to force overiding of any control's colour Fixed many warnings Fixed graph's click event being Y, X and not X, Y

+++ Consor/Controls/Label.cpp
@@ -30,7 +30,7 @@ void Label::ForceResize(const Size& size)
 {
 	size_t width,height;
 
-	_Text = Util::WrapText(_SetText, size.Width, &width, &height);
+	_Text = Util::WrapText(_SetText, (size_t)size.Width, &width, &height);
 	_Size = Size(width, height); //(height > Size.Height ? height : Size.Height));
 
 	OnResize(_Size);
